app/requirements.txt
fastapi
uvicorn
pydantic
pytest
psycopg2-binary
passlib
sqlalchemy
jwt
python-dotenv
pydantic[email]
python-jose[cryptography]

